Completed day24 part1 and part2.
Wrote the interpater, tested it, started running to find the correct input to part1.
Realized that would take days....manually parsed the input program to find both part 1 and part2.
Not a very satisfying solution :(. Might go back later and try to compile/optimize the input program and run inputs in parallel.
Hopefully be able to run the actual serach in a reasonable amount of time.